+++ app/src/main/java/org/juanro/autumandu/data/report/MileageReport.java
@@ -99,8 +99,8 @@ public class MileageReport extends AbstractReport {
             for (BalancedRefueling refueling : refuelings) {
                 if (lastRefuelingMileage > -1) {
                     ZonedDateTime date = ZonedDateTime.ofInstant(refueling.getDate().toInstant(), ZoneId.systemDefault());
-                    float x = date.getYear() * 12 + date.getMonthValue() - 1;
-                    float y = refueling.getMileage() - lastRefuelingMileage;
+                    float x = (float) (date.getYear() * 12 + date.getMonthValue() - 1);
+                    float y = (float) (refueling.getMileage() - lastRefuelingMileage);
 
                     int xIndex = indexOf(x);
                     if (xIndex == -1) {
